deleted laodPaymentSetup file and added function in QueryParamContext
