+++ src/Donations/Micros/PaymentStatus/FailedDonation.tsx
@@ -9,11 +9,10 @@ import { useRouter } from "next/router";
 import { PAYMENT } from "src/Utils/donationStepConstants";
 import InfoIcon from "public/assets/icons/InfoIcon";
 import RetryIcon from "public/assets/icons/RetryIcon";
-import loadPaymentSetup from "src/Utils/loadPaymentSetup";
 
 function FailedDonation({ sendToReturn, donation }: any) {
   const { t } = useTranslation(["common"]);
-  // const [paymentError, setPaymentError] = useState("");
+
   const {
     callbackUrl,
     donationID,
@@ -27,46 +26,15 @@ function FailedDonation({ sendToReturn, donation }: any) {
     setfrequency,
     setdonationStep,
     setcurrency,
-    setshowErrorCard,
-    setpaymentSetup,
-    country,
     paymentError,
     setPaymentError,
     setAmount,
     setcallbackUrl,
     setCallbackMethod,
     setredirectstatus,
-    tenant,
+    loadPaymentSetup,
   } = React.useContext(QueryParamContext);
-  const router = useRouter();
-  const [isPaymentOptionsLoading, setIsPaymentOptionsLoading] =
-    React.useState<boolean>(false);
 
-  // async function loadPaymentSetup(
-  //   projectGUID: string | string[],
-  //   paymentSetupCountry: string
-  // ) {
-  //   setIsPaymentOptionsLoading(true);
-  //   try {
-  //     const requestParams = {
-  //       url: `/app/projects/${projectGUID}/paymentOptions?country=${paymentSetupCountry}`,
-  //       setshowErrorCard,
-  //       tenant,
-  //     };
-  //     const paymentSetupData: any = await apiRequest(requestParams);
-  //     if (paymentSetupData.data) {
-  //       setcurrency(paymentSetupData.data.currency);
-  //       if (!country) {
-  //         setcountry(paymentSetupData.data.effectiveCountry);
-  //       }
-  //       setpaymentSetup(paymentSetupData.data);
-  //     }
-  //     setIsPaymentOptionsLoading(false);
-  //     // setdonationStep(3);
-  //   } catch (err) {
-  //     // console.log(err);
-  //   }
-  // }
   async function getDonation() {
     setIsTaxDeductible(donation.taxDeductionCountry);
     setprojectDetails(donation.project);
@@ -96,14 +64,6 @@ function FailedDonation({ sendToReturn, donation }: any) {
     await loadPaymentSetup({
       projectGUID: donation.project.id,
       paymentSetupCountry: country,
-      setIsPaymentOptionsLoading,
-      setshowErrorCard,
-      tenant,
-      setcurrency,
-      country,
-      setcountry,
-      setpaymentSetup,
-      setprojectDetails,
       shouldSetPaymentDetails: true,
     });
     setcallbackUrl(donation.metadata.callback_url);
